feat: set html lang from locale
