+++ frontend/app/layout.tsx
@@ -1,0 +1,37 @@
+import { cookies } from "next/headers";
+import { Geist, Geist_Mono } from "next/font/google";
+import type { ReactNode } from "react";
+import "./globals.css";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+  preload: false,
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+  preload: false,
+});
+
+export default async function RootLayout({
+  children,
+  params,
+}: Readonly<{ children: ReactNode; params: { locale?: string } }>) {
+  const cookieStore = await cookies();
+  const locale = params.locale ?? cookieStore.get("i18nextLng")?.value ?? "en";
+
+  return (
+    <html lang={locale}>
+      <head>
+        <link rel="icon" href="/favicon.ico" />
+      </head>
+      <body
+        className={`${geistSans.variable} ${geistMono.variable} min-h-screen bg-background font-sans antialiased flex flex-col`}
+      >
+        {children}
+      </body>
+    </html>
+  );
+}
